pyproject.toml
[tool.poetry]
name = "torbot"
version = "3.1.1"
description = "OSINT for the dark web"
authors = ["Akeem King <akeemtlking@gmail.com>"]
license = "GNU GPL"
include = [".env"]

[tool.poetry.dependencies]
python = ">=3.9,<=3.11.4" 
altgraph = "0.17.2"
beautifulsoup4 = "4.11.1"
certifi = "2023.7.22"
charset-normalizer = "2.0.12"
decorator = "5.1.1"
idna = "3.3"
igraph = "0.10.6"
joblib = "1.2.0"
macholib = "1.16"
progress = "1.6"
pyinstaller = "5.13.0"
pyinstaller-hooks-contrib = "2022.7"
PySocks = "1.7.1"
python-dotenv = "0.20.0"
requests = "2.31.0"
requests-mock = "1.9.3"
scikit-learn = "1.3.0"
scipy = "1.10.0"
six = "1.16.0"
sklearn = "0.0"
soupsieve = "2.3.2.post1"
termcolor = "1.1.0"
texttable = "1.6.4"
threadpoolctl = "3.1.0"
threadsafe = "1.0.0"
urllib3 = "1.26.9"
validators = "0.20.0"
yattag = "1.14.0"
treelib = "^1.6.1"
numpy = "1.24.4"
unipath = "^1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
